Tryin to adjust temperature precision.

+++ libraries/TemperatureSensor/TemperatureSensor.h
@@ -16,7 +16,7 @@ public:
 
 private:
   int _sensorPin;
-  double _compute(int RawADC);
+  float _compute(int RawADC);
 };
 
 #endif
